add column, ElementCollection, CollectionTable annotaition in product entity

+++ APNA MARKET/ApnaBazarBackend/src/main/java/com/prashant/apna/bazar/entities/Product.java
@@ -32,7 +32,7 @@ public class Product {
   private double discount;
   private double finalPrice;
   private boolean stock;
-  private Long stockQuantity;
+  private Integer stockQuantity;
 
   @Column(columnDefinition = "TEXT")
   private String description;
